Primer Uso modular, solo falta conexión con la api y crear los endpoints correctos
