+++ js/controllers/controllerFirstUse.js
@@ -1,16 +1,61 @@
-// Controlador principal para el flujo de primer uso
+// Importaciones
+import { 
+    validarPaso1, 
+    inicializarInputsTelefono 
+} from '../utils/validacionesFirstUse.js';
 
-// Variable global que almacena el paso actual
-let pasoActualGlobal = 1;
+import { 
+    guardarDatosPaso1, 
+    restaurarDatosPaso1,
+    guardarDatosPaso2,
+    guardarDatosPaso3,
+    guardarDatosPaso4
+} from '../utils/storageHelperFirstUse.js';
 
-// Función principal para cargar el contenido de cada paso
-function cargarPaso() {
+import { initPaso2, restaurarDatosPaso2 } from './controllerPaso2.js';
+import { initPaso3, restaurarDatosPaso3 } from './controllerPaso3.js';
+import { restaurarDatosPaso4 } from './controllerPaso4.js';
+
+// Variable global
+export let pasoActualGlobal = 1;
+
+// Funciones principales
+export function actualizarIndicadorPaso() {
+    document.querySelectorAll('.paso').forEach(p => p.classList.remove('activo'));
+    const pasos = document.querySelectorAll('.paso');
+    if (pasos[pasoActualGlobal - 1]) {
+        pasos[pasoActualGlobal - 1].classList.add('activo');
+    }
+    console.log("Paso actual: " + pasoActualGlobal);
+}
+
+export function actualizarIndicadorPasoVisual() {
+    document.querySelectorAll('.paso').forEach(p => p.classList.remove('activo'));
+    const pasos = document.querySelectorAll('.paso');
+    if (pasos[pasoActualGlobal - 1]) {
+        pasos[pasoActualGlobal - 1].classList.add('activo');
+    }
+    console.log("Indicador actualizado para el Paso: " + pasoActualGlobal);
+}
+
+export function limpiarFormulario() {
+    const contenedor = document.getElementById("contenido-dinamico");
+    if (contenedor) {
+        const formularios = contenedor.getElementsByTagName('form');
+        for (let i = 0; i < formularios.length; i++) {
+            formularios[i].reset();
+        }
+    } else {
+        console.error(`Contenedor con ID 'contenido-dinamico' no encontrado para limpiar formularios.`);
+    }
+}
+
+export function cargarPaso() {
     fetch(`pasosPrimerUso/paso${pasoActualGlobal}.html`)
         .then(res => res.text())
         .then(html => {
             document.getElementById("contenido-dinamico").innerHTML = html;
             document.getElementById("paso-actual").textContent = pasoActualGlobal;
-
             actualizarIndicadorPasoVisual();
 
             setTimeout(() => {
@@ -42,31 +87,7 @@ function cargarPaso() {
     }
 }
 
-// Función para actualizar el indicador visual del paso actual
-function actualizarIndicadorPasoVisual() {
-    document.querySelectorAll('.paso').forEach(p => p.classList.remove('activo'));
-    const pasos = document.querySelectorAll('.paso');
-    if (pasos[pasoActualGlobal - 1]) {
-        pasos[pasoActualGlobal - 1].classList.add('activo');
-    }
-    console.log("Indicador actualizado para el Paso: " + pasoActualGlobal);
-}
-
-// Función para limpiar formularios
-function limpiarFormulario() {
-    const contenedor = document.getElementById("contenido-dinamico");
-    if (contenedor) {
-        const formularios = contenedor.getElementsByTagName('form');
-        for (let i = 0; i < formularios.length; i++) {
-            formularios[i].reset();
-        }
-    } else {
-        console.error(`Contenedor con ID 'contenido-dinamico' no encontrado para limpiar formularios.`);
-    }
-}
-
-// Función para avanzar al siguiente paso
-function siguientePaso() {
+export function siguientePaso() {
     if (pasoActualGlobal === 1) {
         if (!validarPaso1()) return;
         guardarDatosPaso1();
@@ -88,22 +109,19 @@ function siguientePaso() {
             Swal.fire({
                 icon: 'success',
                 title: '¡Enhorabuena!',
-                html: `
-          <p>Haz creado tu propio equipo.</p>
-          <p>puedes dirigirte al panel principal para gestionarlo</p>
-        `,
+                html: `<p>Haz creado tu propio equipo.</p><p>puedes dirigirte al panel principal para gestionarlo</p>`,
                 confirmButtonText: 'Ir al Inicio',
                 confirmButtonColor: '#28a745'
             }).then(() => {
                 window.location.href = './PlataformaWebInicio/PW_Inicio.html';
             });
         });
-
         return;
     }
 
     if (pasoActualGlobal === 2) guardarDatosPaso2();
     if (pasoActualGlobal === 3) guardarDatosPaso3();
+    if (pasoActualGlobal === 4 && typeof guardarDatosPaso4 === 'function') guardarDatosPaso4();
 
     if (pasoActualGlobal < 4) {
         pasoActualGlobal++;
@@ -111,16 +129,14 @@ function siguientePaso() {
     }
 }
 
-// Función para retroceder al paso anterior
-function anteriorPaso() {
+export function anteriorPaso() {
     if (pasoActualGlobal > 1) {
         pasoActualGlobal--;
         cargarPaso();
     }
 }
 
-// Función para cancelar el proceso
-function cancelarPaso() {
+export function cancelarPaso() {
     const mensaje = "Si cancelás ahora, se perderán los datos ingresados.";
 
     if (typeof Swal !== 'undefined') {
@@ -153,8 +169,7 @@ function cancelarPaso() {
     }
 }
 
-// Función para navegar a un paso específico
-function navegarAPaso(numeroPasoDestino) {
+export function navegarAPaso(numeroPasoDestino) {
     const pasoInt = parseInt(numeroPasoDestino);
 
     if (pasoInt < 1 || pasoInt > 4) {
@@ -190,48 +205,65 @@ function navegarAPaso(numeroPasoDestino) {
         });
 }
 
-// Inicializar componentes específicos de cada paso
-function inicializarComponentesPaso(pasoActualGlobal) {
-    setTimeout(() => {
+export function inicializarComponentesPaso(pasoActualGlobal) {
+  fetch(`pasosPrimerUso/paso${pasoActualGlobal}.html`)
+    .then(res => res.text())
+    .then(html => {
+      document.getElementById("contenido-dinamico").innerHTML = html;
+      document.getElementById("paso-actual").textContent = pasoActualGlobal;
+      actualizarIndicadorPaso();
+
+      setTimeout(() => {
         inicializarInputsTelefono();
 
-        requestAnimationFrame(() => {
-            if (pasoActualGlobal === 1) {
-                restaurarDatosPaso1();
-            }
-            if (pasoActualGlobal === 2) {
-                initPaso2();
-                setTimeout(() => {
-                    restaurarDatosPaso2();
-                }, 500);
-            }
-            if (pasoActualGlobal === 3) {
-                initPaso3();
-            }
-            if (pasoActualGlobal === 4) {
-                restaurarDatosPaso4();
-            }
+        requestAnimationFrame(async () => {
+          if (pasoActualGlobal === 1) {
+            restaurarDatosPaso1();
+          }
+
+          if (pasoActualGlobal === 2) {
+            initPaso2();
+            setTimeout(() => {
+              restaurarDatosPaso2();
+            }, 500);
+          }
+
+          if (pasoActualGlobal === 3) {
+            await initPaso3(); // Asegúrate de que initPaso3 sea async
+            const { listaTecnicos } = await import('./controllerPaso3.js');
+            restaurarDatosPaso4(listaTecnicos);
+          }
+
+          if (pasoActualGlobal === 4) {
+            const { listaTecnicos } = await import('./controllerPaso3.js');
+            restaurarDatosPaso4(listaTecnicos);
+          }
         });
-    }, 0);
+      }, 0);
+
+      const btnAtras = document.getElementById("btn-atras");
+      if (btnAtras) {
+        btnAtras.style.display = pasoActualGlobal === 1 ? "none" : "inline-flex";
+      }
+    });
 }
 
-// Carga inicial
-document.addEventListener('DOMContentLoaded', () => {
+export function inicializarAplicacion() {
     cargarPaso();
-});
+    
+    // Configurar event listeners para navegación entre pasos
+    document.addEventListener("click", function (e) {
+        const botonNavegar = e.target.closest("[data-navegar-paso]");
+        if (botonNavegar) {
+            const pasoDestino = botonNavegar.dataset.navegarPaso;
+            navegarAPaso(pasoDestino);
+            return;
+        }
 
-// Event listener para botones de navegación
-document.addEventListener("click", function (e) {
-    const botonNavegar = e.target.closest("[data-navegar-paso]");
-    if (botonNavegar) {
-        const pasoDestino = botonNavegar.dataset.navegarPaso;
-        navegarAPaso(pasoDestino);
-        return;
-    }
-
-    const boton = e.target.closest("[data-paso]");
-    if (boton) {
-        const pasoDestino = boton.dataset.paso;
-        navegarAPaso(pasoDestino);
-    }
-});
+        const boton = e.target.closest("[data-paso]");
+        if (boton) {
+            const pasoDestino = boton.dataset.paso;
+            navegarAPaso(pasoDestino);
+        }
+    });
+}
