+++ pasosPrimerUso/paso1.html
@@ -208,8 +208,8 @@
     <!-- SweetAlert -->
     <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
 
-    <!-- Tu script personalizado -->
-    <script src="js/scriptPasos.js"></script>
+    <!-- Script personalizado -->
+    <script type="module" src="../js/scriptPasos.js"></script>
 
     <span id="paso-actual">1</span>
 </body>
