+++ pasosPrimerUso/paso2.html
@@ -201,7 +201,8 @@
 
     <!-- SweetAlert -->
     <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
-    <script src="js/scriptPasos.js" defer></script>
+    <!-- Script personalizado -->
+    <script type="module" src="../js/scriptPasos.js" defer></script>
     </div>
     </div>
     <span id="paso-actual">2</span>
