+++ pasosPrimerUso/paso3.html
@@ -269,7 +269,8 @@
 
     <!-- SweetAlert -->
     <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
-    <script src="js/scriptPasos.js" defer></script>
+    <!-- Script personalizado -->
+    <script type="module" src="../js/scriptPasos.js" defer></script>
     </div>
     </div>
     <span id="paso-actual">3</span>
