+++ pasosPrimerUso/paso4.html
@@ -203,8 +203,9 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap/5.3.0/js/bootstrap.bundle.min.js"></script>
 
     <!-- JS -->
-    <script src="js/scriptPasos.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
+    <!-- Script personalizado -->
+    <script type="module" src="../js/scriptPasos.js" defer></script>
     </div>
     </div>
     <span id="paso-actual">4</span>
