+++ primerUsoPasos.html
@@ -11,6 +11,7 @@
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100..900;1,100..900&display=swap"
         rel="stylesheet">
+    <link rel="icon" href="favicon.ico">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css">
     <link rel="stylesheet" href="css/stylePrimerUsoPasos.css">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/intl-tel-input@17.0.19/build/css/intlTelInput.min.css" />
@@ -175,7 +176,7 @@
                             </svg>
                             Atrás</button>
                         <button class="cancel-button" onclick="cancelarPaso()">Cancelar</button>
-                        <button type="button" class="next-button" onclick="accionSiguientePaso()">Siguiente</button>
+                        <button type="button" class="next-button" onclick="siguientePaso()">Siguiente</button>
                     </div>
                 </footer>
 
@@ -183,11 +184,7 @@
                 <script src="https://cdn.jsdelivr.net/npm/intl-tel-input@17.0.19/build/js/intlTelInput.min.js"></script>
                 <script src="https://cdn.jsdelivr.net/npm/intl-tel-input@17.0.19/build/js/utils.js"></script>
                 <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
-                <script src="/js/utils/storageHelperFirstUse.js"></script>
-                <script src="/js/utils/validacionesFirstUse.js"></script>
-                <script src="/js/services/serviceFirstUse.js"></script>
-                <script src="/js/controllers/controllerFirstUse.js"></script>
-                <script src="/js/scriptPasos.js"></script>
+                <script type="module" src="./js/scriptPasos.js"></script>
             </div>
         </div>
     </div>
